requirements.txt
# GUI
PyQt5>=5.15.10

# Аудио
sounddevice>=0.4.6
numpy>=1.24.0

# Speech Recognition
faster-whisper>=0.10.0
vosk>=0.3.45
torch>=2.0.0
torchaudio>=2.0.0

# AI/LLM
requests>=2.31.0
google-generativeai>=0.3.2

# Утилиты
PyYAML>=6.0.1
colorlog>=6.8.0